requirements.txt
# Flask y dependencias web
Flask>=3.0.0
Werkzeug>=3.0.1

# Manejo de peticiones HTTP
requests>=2.31.0

# Variables de entorno
python-dotenv>=1.0.0

# Base de datos para producción (Railway)
psycopg2-binary>=2.9.7